semantic_mcp_server/requirements.txt
# MCP Protocol
fastmcp>=2.0.0,<3.0.0
mcp>=1.0.0,<2.0.0

# Neo4j only — no Postgres/Qdrant in semantic MCP
neo4j>=5.19.0,<6.0.0

# HTTP / ASGI
httpx>=0.27.0,<1.0.0
starlette>=0.37.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
websockets>=13.0,<14.0

# Utilities
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Auth
python-jose[cryptography]>=3.3.0

prometheus-client>=0.21.0
structlog>=24.4.0
